Improve link input dialog appearance (#9)
This PR adds some small improvements to the link insertion dialog. I
originally considered making the dialog fully customizable, but the
changes I wanted to make (to align it with Thunder's design) were so
minimal I figured it wouldn't be worth it. In fact, most of the changes
came down to *removing* custom styling and using built-in properties of
`AlertDialog`.

### Before


https://github.com/thunder-app/markdown-editor/assets/7417301/03a38c68-e5c5-4834-b0bc-cb280091b71e

### After


https://github.com/thunder-app/markdown-editor/assets/7417301/daf16c8c-c178-41b3-aa7e-54bef3943b1c